ajuste em migrations

+++ database/migrations/2022_01_29_010338_currency_trade_change_currency_rate_add.php
@@ -14,7 +14,7 @@ class CurrencyTradeChangeCurrencyRateAdd extends Migration
     public function up()
     {
         Schema::table('currency_trades', function (Blueprint $table) {
-            $table->decimal('currency_rate', 8, 5);
+            $table->double('currency_rate', 12, 4);
         });    
     }
